# fix: fix threads freeze on waiting state when logging request body through custom logger