## Situation
In pull request #446, a new requests logger was introduced, which logs requests separately from ktor client. The custom logger intercepts the [`HttpSendPipeline`](https://api.ktor.io/ktor-client/ktor-client-core/io.ktor.client.request/-http-send-pipeline/index.html) on [`Monitoring Phase`](https://api.ktor.io/ktor-client/ktor-client-core/io.ktor.client.request/-http-send-pipeline/-phases/-monitoring.html). 

For each request intercepted, an [`HttpRequestBuilder`](https://api.ktor.io/ktor-client/ktor-client-core/io.ktor.client.request/-http-request-builder/index.html) instance is available for read operations. The `HttpRequestBuilder` allows access for a few attributes request such as [`body`](https://api.ktor.io/ktor-client/ktor-client-core/io.ktor.client.call/body.html) and [`headers`](https://api.ktor.io/ktor-client/ktor-client-core/io.ktor.client.request/-http-request-builder/headers.html). 

Since an interceptor would not necessarily attempt to read the `body` attribute, its type, based on its size, can be one of two types:
1) `String`: Default type.
2) [`OutputStreamContent`](https://api.ktor.io/ktor-http/io.ktor.http.content/-output-stream-content/index.html): In the case of large request bodies, which can be costly to read and transform into a `String` type.

In the case of large request bodies, we attempt to write the data exposed by the `OutputStreamContent` body into a manually-managed [`ByteChannel`](https://api.ktor.io/ktor-io/io.ktor.utils.io/-byte-channel/index.html) instance, read the bytes written into the `ByteChannel`, and decode the bytes into a `utf-8` text afterwards. 

```kotlin
with(ByteChannel()) {
    body.writeTo(this)
    return readRemaining().readText()
}
```

The [`writeTo`](https://api.ktor.io/ktor-http/io.ktor.http.content/-output-stream-content/write-to.html) method, which us used in the previous implementation to write data into the `ByteChannel` is run inside a [`withBlocking`](https://github.com/ktorio/ktor/blob/6763a7344059c1f9dfdb2f0816e98559862ef604/ktor-http/jvm/src/io/ktor/http/content/BlockingBridge.kt#L26) scope.

For some reason, when the write operation is triggered, at some point the writing execution thread, along with all other threads, go into a waiting state. Please check the following [thread dump](https://www.perfmatrix.com/thread-dump-overview/) below, which is generated by [VisualVM](https://visualvm.github.io/), in order to have a clear image of threads state after the writing operation is triggered. Based on my own observations and opinion, the threads dump is indicating a state somewhat similar to a [deadlock](https://en.wikipedia.org/wiki/Deadlock).
<details>
<summary><b>Threads Dump</b></summary>

    2024-04-02 22:01:51 
    Full thread dump OpenJDK 64-Bit Server VM (18.0.2.1+1-1 mixed mode, sharing):
    
    Threads class SMR info:
    _java_thread_list=0x00006000007eeea0, length=29, elements={
    0x000000014d809800, 0x000000012c808200, 0x000000014d809000, 0x000000012c809e00,
    0x000000014d010000, 0x000000014d010600, 0x000000012c808e00, 0x000000014d80ae00,
    0x000000014d819600, 0x000000014d81a600, 0x000000012c80a400, 0x000000014d819000,
    0x000000013d830600, 0x000000014d0c7e00, 0x000000012c80aa00, 0x000000012c8d9e00,
    0x000000012d185400, 0x000000013d0b4a00, 0x000000014d0dca00, 0x000000013d11e400,
    0x000000013d11ea00, 0x000000013e825400, 0x000000012f846000, 0x000000014d9c9c00,
    0x000000012c8cd600, 0x000000013d83ae00, 0x000000014d8eca00, 0x000000014e825600,
    0x000000012c8ce200
    }
    
    "main" #1 prio=5 os_prio=31 cpu=361.96ms elapsed=6.98s tid=0x000000014d809800 nid=8963 waiting on condition  [0x000000016b9be000]
       java.lang.Thread.State: WAITING (parking)
            at jdk.internal.misc.Unsafe.park(java.base@18.0.2.1/Native Method)
            - parking to wait for  <0x0000000700400000> (a java.util.concurrent.CompletableFuture$Signaller)
            at java.util.concurrent.locks.LockSupport.park(java.base@18.0.2.1/LockSupport.java:211)
            at java.util.concurrent.CompletableFuture$Signaller.block(java.base@18.0.2.1/CompletableFuture.java:1864)
            at java.util.concurrent.ForkJoinPool.unmanagedBlock(java.base@18.0.2.1/ForkJoinPool.java:3464)
            at java.util.concurrent.ForkJoinPool.managedBlock(java.base@18.0.2.1/ForkJoinPool.java:3435)
            at java.util.concurrent.CompletableFuture.waitingGet(java.base@18.0.2.1/CompletableFuture.java:1898)
            at java.util.concurrent.CompletableFuture.get(java.base@18.0.2.1/CompletableFuture.java:2072)
            at com.expediagroup.sdk.fraudpreventionv2.client.FraudPreventionV2Client.screenOrderWithResponse(FraudPreventionV2Client.kt:408)
            at com.expediagroup.sdk.fraudpreventionv2.client.FraudPreventionV2Client.screenOrder(FraudPreventionV2Client.kt:374)
            at org.jarrah.Main.main(Main.java:33)
    
       Locked ownable synchronizers:
            - None
    
    "Reference Handler" #2 daemon prio=10 os_prio=31 cpu=0.50ms elapsed=6.97s tid=0x000000012c808200 nid=17667 waiting on condition  [0x000000016c812000]
       java.lang.Thread.State: RUNNABLE
            at java.lang.ref.Reference.waitForReferencePendingList(java.base@18.0.2.1/Native Method)
            at java.lang.ref.Reference.processPendingReferences(java.base@18.0.2.1/Reference.java:253)
            at java.lang.ref.Reference$ReferenceHandler.run(java.base@18.0.2.1/Reference.java:215)
    
       Locked ownable synchronizers:
            - None
    
    "Finalizer" #3 daemon prio=8 os_prio=31 cpu=0.11ms elapsed=6.97s tid=0x000000014d809000 nid=17923 in Object.wait()  [0x000000016ca1e000]
       java.lang.Thread.State: WAITING (on object monitor)
            at java.lang.Object.wait(java.base@18.0.2.1/Native Method)
            - waiting on <0x0000000700001b18> (a java.lang.ref.ReferenceQueue$Lock)
            at java.lang.ref.ReferenceQueue.remove(java.base@18.0.2.1/ReferenceQueue.java:155)
            - locked <0x0000000700001b18> (a java.lang.ref.ReferenceQueue$Lock)
            at java.lang.ref.ReferenceQueue.remove(java.base@18.0.2.1/ReferenceQueue.java:176)
            at java.lang.ref.Finalizer$FinalizerThread.run(java.base@18.0.2.1/Finalizer.java:183)
    
       Locked ownable synchronizers:
            - None
    
    "Signal Dispatcher" #4 daemon prio=9 os_prio=31 cpu=0.43ms elapsed=6.96s tid=0x000000012c809e00 nid=24067 waiting on condition  [0x0000000000000000]
       java.lang.Thread.State: RUNNABLE
    
       Locked ownable synchronizers:
            - None
    
    "Service Thread" #5 daemon prio=9 os_prio=31 cpu=0.18ms elapsed=6.96s tid=0x000000014d010000 nid=31747 runnable  [0x0000000000000000]
       java.lang.Thread.State: RUNNABLE
    
       Locked ownable synchronizers:
            - None
    
    "Monitor Deflation Thread" #6 daemon prio=9 os_prio=31 cpu=0.09ms elapsed=6.96s tid=0x000000014d010600 nid=24323 runnable  [0x0000000000000000]
       java.lang.Thread.State: RUNNABLE
    
       Locked ownable synchronizers:
            - None
    
    "C2 CompilerThread0" #7 daemon prio=9 os_prio=31 cpu=1230.54ms elapsed=6.96s tid=0x000000012c808e00 nid=24835 waiting on condition  [0x0000000000000000]
       java.lang.Thread.State: RUNNABLE
       No compile task
    
       Locked ownable synchronizers:
            - None
    
    "C1 CompilerThread0" #10 daemon prio=9 os_prio=31 cpu=610.84ms elapsed=6.96s tid=0x000000014d80ae00 nid=31235 waiting on condition  [0x0000000000000000]
       java.lang.Thread.State: RUNNABLE
       No compile task
    
       Locked ownable synchronizers:
            - None
    
    "Sweeper thread" #11 daemon prio=9 os_prio=31 cpu=17.59ms elapsed=6.96s tid=0x000000014d819600 nid=30979 runnable  [0x0000000000000000]
       java.lang.Thread.State: RUNNABLE
    
       Locked ownable synchronizers:
            - None
    
    "Common-Cleaner" #12 daemon prio=8 os_prio=31 cpu=0.50ms elapsed=6.96s tid=0x000000014d81a600 nid=30467 in Object.wait()  [0x000000016d98a000]
       java.lang.Thread.State: TIMED_WAITING (on object monitor)
            at java.lang.Object.wait(java.base@18.0.2.1/Native Method)
            - waiting on <0x00000007000014d8> (a java.lang.ref.ReferenceQueue$Lock)
            at java.lang.ref.ReferenceQueue.remove(java.base@18.0.2.1/ReferenceQueue.java:155)
            - locked <0x00000007000014d8> (a java.lang.ref.ReferenceQueue$Lock)
            at jdk.internal.ref.CleanerImpl.run(java.base@18.0.2.1/CleanerImpl.java:140)
            at java.lang.Thread.run(java.base@18.0.2.1/Thread.java:833)
            at jdk.internal.misc.InnocuousThread.run(java.base@18.0.2.1/InnocuousThread.java:162)
    
       Locked ownable synchronizers:
            - None
    
    "Monitor Ctrl-Break" #13 daemon prio=5 os_prio=31 cpu=7.63ms elapsed=6.86s tid=0x000000012c80a400 nid=30211 runnable  [0x000000016db96000]
       java.lang.Thread.State: RUNNABLE
            at sun.nio.ch.SocketDispatcher.read0(java.base@18.0.2.1/Native Method)
            at sun.nio.ch.SocketDispatcher.read(java.base@18.0.2.1/SocketDispatcher.java:47)
            at sun.nio.ch.NioSocketImpl.tryRead(java.base@18.0.2.1/NioSocketImpl.java:258)
            at sun.nio.ch.NioSocketImpl.implRead(java.base@18.0.2.1/NioSocketImpl.java:309)
            at sun.nio.ch.NioSocketImpl.read(java.base@18.0.2.1/NioSocketImpl.java:347)
            at sun.nio.ch.NioSocketImpl$1.read(java.base@18.0.2.1/NioSocketImpl.java:800)
            at java.net.Socket$SocketInputStream.read(java.base@18.0.2.1/Unknown Source)
            at sun.nio.cs.StreamDecoder.readBytes(java.base@18.0.2.1/StreamDecoder.java:270)
            at sun.nio.cs.StreamDecoder.implRead(java.base@18.0.2.1/StreamDecoder.java:313)
            at sun.nio.cs.StreamDecoder.read(java.base@18.0.2.1/StreamDecoder.java:188)
            - locked <0x0000000700002880> (a java.io.InputStreamReader)
            at java.io.InputStreamReader.read(java.base@18.0.2.1/InputStreamReader.java:176)
            at java.io.BufferedReader.fill(java.base@18.0.2.1/BufferedReader.java:162)
            at java.io.BufferedReader.readLine(java.base@18.0.2.1/BufferedReader.java:329)
            - locked <0x0000000700002880> (a java.io.InputStreamReader)
            at java.io.BufferedReader.readLine(java.base@18.0.2.1/BufferedReader.java:396)
            at com.intellij.rt.execution.application.AppMainV2$1.run(AppMainV2.java:53)
    
       Locked ownable synchronizers:
            - <0x0000000700a3ede0> (a java.util.concurrent.locks.ReentrantLock$NonfairSync)
    
    "Notification Thread" #14 daemon prio=9 os_prio=31 cpu=0.03ms elapsed=6.86s tid=0x000000014d819000 nid=25859 runnable  [0x0000000000000000]
       java.lang.Thread.State: RUNNABLE
    
       Locked ownable synchronizers:
            - None
    
    "DefaultDispatcher-worker-1" #15 daemon prio=5 os_prio=31 cpu=420.33ms elapsed=6.71s tid=0x000000013d830600 nid=28163 waiting on condition  [0x000000016ebf2000]
       java.lang.Thread.State: TIMED_WAITING (parking)
            at jdk.internal.misc.Unsafe.park(java.base@18.0.2.1/Native Method)
            at java.util.concurrent.locks.LockSupport.parkNanos(java.base@18.0.2.1/LockSupport.java:376)
            at io.ktor.utils.io.jvm.javaio.DefaultParking.park(Pollers.kt:44)
            at io.ktor.utils.io.jvm.javaio.BlockingAdapter.parkingLoop(Blocking.kt:268)
            at io.ktor.utils.io.jvm.javaio.BlockingAdapter.submitAndAwait(Blocking.kt:233)
            at io.ktor.utils.io.jvm.javaio.BlockingAdapter.submitAndAwait(Blocking.kt:201)
            at io.ktor.utils.io.jvm.javaio.OutputAdapter.write(Blocking.kt:120)
            - locked <0x0000000700cc4d38> (a io.ktor.utils.io.jvm.javaio.OutputAdapter)
            at com.fasterxml.jackson.core.json.UTF8JsonGenerator._flushBuffer(UTF8JsonGenerator.java:2208)
            at com.fasterxml.jackson.core.json.UTF8JsonGenerator._writeStringSegments(UTF8JsonGenerator.java:1348)
            at com.fasterxml.jackson.core.json.UTF8JsonGenerator.writeString(UTF8JsonGenerator.java:522)
            at com.fasterxml.jackson.databind.ser.std.StringSerializer.serialize(StringSerializer.java:41)
            at com.fasterxml.jackson.databind.ser.BeanPropertyWriter.serializeAsField(BeanPropertyWriter.java:732)
            at com.fasterxml.jackson.databind.ser.std.BeanSerializerBase.serializeFields(BeanSerializerBase.java:770)
            at com.fasterxml.jackson.databind.ser.BeanSerializer.serialize(BeanSerializer.java:183)
            at com.fasterxml.jackson.databind.ser.BeanPropertyWriter.serializeAsField(BeanPropertyWriter.java:732)
            at com.fasterxml.jackson.databind.ser.std.BeanSerializerBase.serializeFields(BeanSerializerBase.java:770)
            at com.fasterxml.jackson.databind.ser.BeanSerializer.serialize(BeanSerializer.java:183)
            at com.fasterxml.jackson.databind.ser.BeanPropertyWriter.serializeAsField(BeanPropertyWriter.java:732)
            at com.fasterxml.jackson.databind.ser.std.BeanSerializerBase.serializeFields(BeanSerializerBase.java:770)
            at com.fasterxml.jackson.databind.ser.BeanSerializer.serialize(BeanSerializer.java:183)
            at com.fasterxml.jackson.databind.ser.DefaultSerializerProvider._serialize(DefaultSerializerProvider.java:502)
            at com.fasterxml.jackson.databind.ser.DefaultSerializerProvider.serializeValue(DefaultSerializerProvider.java:341)
            at com.fasterxml.jackson.databind.ObjectMapper._writeValueAndClose(ObjectMapper.java:4793)
            at com.fasterxml.jackson.databind.ObjectMapper.writeValue(ObjectMapper.java:3997)
            at io.ktor.serialization.jackson.JacksonConverter$serializeNullable$2.invokeSuspend(JacksonConverter.kt:81)
            at io.ktor.serialization.jackson.JacksonConverter$serializeNullable$2.invoke(JacksonConverter.kt)
            at io.ktor.serialization.jackson.JacksonConverter$serializeNullable$2.invoke(JacksonConverter.kt)
            at io.ktor.http.content.OutputStreamContent$writeTo$2.invokeSuspend(OutputStreamContent.kt:28)
            at io.ktor.http.content.OutputStreamContent$writeTo$2.invoke(OutputStreamContent.kt)
            at io.ktor.http.content.OutputStreamContent$writeTo$2.invoke(OutputStreamContent.kt)
            at io.ktor.http.content.BlockingBridgeKt.withBlocking(BlockingBridge.kt:28)
            at io.ktor.http.content.OutputStreamContent.writeTo(OutputStreamContent.kt:24)
            at com.expediagroup.sdk.core.plugin.logging.RequestBodyLogger$Plugin.getBody(RequestBodyLogger.kt:53)
            at com.expediagroup.sdk.core.plugin.logging.RequestBodyLogger$Plugin.access$getBody(RequestBodyLogger.kt:33)
            at com.expediagroup.sdk.core.plugin.logging.RequestBodyLogger$Plugin$install$1.invokeSuspend(RequestBodyLogger.kt:42)
            at com.expediagroup.sdk.core.plugin.logging.RequestBodyLogger$Plugin$install$1.invoke(RequestBodyLogger.kt)
            at com.expediagroup.sdk.core.plugin.logging.RequestBodyLogger$Plugin$install$1.invoke(RequestBodyLogger.kt)
            at io.ktor.util.pipeline.SuspendFunctionGun.loop(SuspendFunctionGun.kt:120)
            at io.ktor.util.pipeline.SuspendFunctionGun.proceed(SuspendFunctionGun.kt:78)
            at io.ktor.util.pipeline.SuspendFunctionGun.proceedWith(SuspendFunctionGun.kt:88)
            at io.ktor.client.plugins.logging.Logging$setupRequestLogging$1.invokeSuspend(Logging.kt:90)
            at io.ktor.client.plugins.logging.Logging$setupRequestLogging$1.invoke(Logging.kt)
            at io.ktor.client.plugins.logging.Logging$setupRequestLogging$1.invoke(Logging.kt)
            at io.ktor.util.pipeline.SuspendFunctionGun.loop(SuspendFunctionGun.kt:120)
            at io.ktor.util.pipeline.SuspendFunctionGun.proceed(SuspendFunctionGun.kt:78)
            at io.ktor.util.pipeline.SuspendFunctionGun.execute$ktor_utils(SuspendFunctionGun.kt:98)
            at io.ktor.util.pipeline.Pipeline.execute(Pipeline.kt:77)
            at io.ktor.client.plugins.HttpSend$DefaultSender.execute(HttpSend.kt:138)
            at com.expediagroup.sdk.core.plugin.authentication.AuthenticationHookBuilder$build$1.invokeSuspend(AuthenticationHookFactory.kt:60)
            at com.expediagroup.sdk.core.plugin.authentication.AuthenticationHookBuilder$build$1.invoke(AuthenticationHookFactory.kt)
            at com.expediagroup.sdk.core.plugin.authentication.AuthenticationHookBuilder$build$1.invoke(AuthenticationHookFactory.kt)
            at io.ktor.client.plugins.HttpSend$InterceptedSender.execute(HttpSend.kt:116)
            at io.ktor.client.plugins.HttpTimeout$Plugin$install$1.invokeSuspend(HttpTimeout.kt:174)
            at io.ktor.client.plugins.HttpTimeout$Plugin$install$1.invoke(HttpTimeout.kt)
            at io.ktor.client.plugins.HttpTimeout$Plugin$install$1.invoke(HttpTimeout.kt)
            at io.ktor.client.plugins.HttpSend$InterceptedSender.execute(HttpSend.kt:116)
            at io.ktor.client.plugins.auth.Auth$Plugin$install$2.invokeSuspend(Auth.kt:67)
            at io.ktor.client.plugins.auth.Auth$Plugin$install$2.invoke(Auth.kt)
            at io.ktor.client.plugins.auth.Auth$Plugin$install$2.invoke(Auth.kt)
            at io.ktor.client.plugins.HttpSend$InterceptedSender.execute(HttpSend.kt:116)
            at io.ktor.client.plugins.HttpRedirect$Plugin$install$1.invokeSuspend(HttpRedirect.kt:64)
            at io.ktor.client.plugins.HttpRedirect$Plugin$install$1.invoke(HttpRedirect.kt)
            at io.ktor.client.plugins.HttpRedirect$Plugin$install$1.invoke(HttpRedirect.kt)
            at io.ktor.client.plugins.HttpSend$InterceptedSender.execute(HttpSend.kt:116)
            at io.ktor.client.plugins.HttpCallValidator$Companion$install$3.invokeSuspend(HttpCallValidator.kt:151)
            at io.ktor.client.plugins.HttpCallValidator$Companion$install$3.invoke(HttpCallValidator.kt)
            at io.ktor.client.plugins.HttpCallValidator$Companion$install$3.invoke(HttpCallValidator.kt)
            at io.ktor.client.plugins.HttpSend$InterceptedSender.execute(HttpSend.kt:116)
            at io.ktor.client.plugins.HttpSend$Plugin$install$1.invokeSuspend(HttpSend.kt:104)
            at io.ktor.client.plugins.HttpSend$Plugin$install$1.invoke(HttpSend.kt)
            at io.ktor.client.plugins.HttpSend$Plugin$install$1.invoke(HttpSend.kt)
            at io.ktor.util.pipeline.SuspendFunctionGun.loop(SuspendFunctionGun.kt:120)
            at io.ktor.util.pipeline.SuspendFunctionGun.proceed(SuspendFunctionGun.kt:78)
            at io.ktor.util.pipeline.SuspendFunctionGun.proceedWith(SuspendFunctionGun.kt:88)
            at io.ktor.client.plugins.DefaultTransformKt$defaultTransformers$1.invokeSuspend(DefaultTransform.kt:57)
            at io.ktor.client.plugins.DefaultTransformKt$defaultTransformers$1.invoke(DefaultTransform.kt)
            at io.ktor.client.plugins.DefaultTransformKt$defaultTransformers$1.invoke(DefaultTransform.kt)
            at io.ktor.util.pipeline.SuspendFunctionGun.loop(SuspendFunctionGun.kt:120)
            at io.ktor.util.pipeline.SuspendFunctionGun.proceed(SuspendFunctionGun.kt:78)
            at io.ktor.util.pipeline.SuspendFunctionGun.proceedWith(SuspendFunctionGun.kt:88)
            at io.ktor.client.plugins.contentnegotiation.ContentNegotiation$Plugin$install$1.invokeSuspend(ContentNegotiation.kt:252)
            at io.ktor.client.plugins.contentnegotiation.ContentNegotiation$Plugin$install$1.invoke(ContentNegotiation.kt)
            at io.ktor.client.plugins.contentnegotiation.ContentNegotiation$Plugin$install$1.invoke(ContentNegotiation.kt)
            at io.ktor.util.pipeline.SuspendFunctionGun.loop(SuspendFunctionGun.kt:120)
            at io.ktor.util.pipeline.SuspendFunctionGun.proceed(SuspendFunctionGun.kt:78)
…