@OmarAlJarrah 2024-04-23
## Sample Request Data
[Here](https://github.com/ExpediaGroup/expediagroup-java-sdk/files/15079869/body.json) is a sample request used for testing the new implementation.